Move TJA parsing funcs to `parsers.py`
