+++ src/tja2fumen/parsers.py
@@ -4,7 +4,7 @@ import re
 from tja2fumen.utils import readStruct, getBool, shortHex
 from tja2fumen.constants import (
     # TJA constants
-    HEADER_GLOBAL, HEADER_COURSE, BRANCH_COMMANDS, MEASURE_COMMANDS, COMMAND, NORMALIZE_COURSE,
+    HEADER_GLOBAL, HEADER_COURSE, BRANCH_COMMANDS, MEASURE_COMMANDS, COMMAND, NORMALIZE_COURSE, TJA_NOTE_TYPES,
     # Fumen constants
     branchNames, noteTypes
 )
@@ -51,7 +51,9 @@ def parseTJA(fnameTJA):
     songs = {}
     for courseName, courseLines in courses.items():
         courseHeader, courseMeasures = getCourse(headers, courseLines)
-        songs[courseName] = applyFumenStructureToParsedTJA(headers, courseHeader, courseMeasures)
+        tja = applyFumenStructureToParsedTJA(headers, courseHeader, courseMeasures)
+        tja['measures'] = preprocessTJAMeasures(tja)
+        songs[courseName] = tja
 
     return songs
 
@@ -267,6 +269,95 @@ def applyFumenStructureToParsedTJA(globalHeader, courseHeader, measures):
         song['measures'].append(measure)
 
     return song
+
+
+def preprocessTJAMeasures(tja):
+    """
+    Merge TJA 'data' and 'event' fields into a single measure property, and split
+    measures into sub-measures whenever a mid-measure BPM change occurs.
+
+    The TJA parser produces measure objects with two important properties:
+      - 'data': Contains the note data (1: don, 2: ka, etc.) along with spacing (s)
+      - 'events' Contains event commands such as MEASURE, BPMCHANGE, GOGOTIME, etc.
+
+    However, notes and events can be intertwined within a single measure. So, it's
+    not possible to process them separately; they must be considered as single sequence.
+
+    A particular danger is BPM changes. TJA allows multiple BPMs within a single measure,
+    but the fumen format permits one BPM per measure. So, a TJA measure must be split up
+    if it has multiple BPM changes within a measure.
+
+    In the future, this logic should probably be moved into the TJA parser itself.
+    """
+    currentBPM = 0
+    currentScroll = 1.0
+    currentGogo = False
+    currentBarline = True
+
+    measuresCorrected = []
+    for measure in tja['measures']:
+        # Step 1: Combine notes and events
+        notes = [{'pos': i, 'type': 'note', 'value': TJA_NOTE_TYPES[note]}
+                 for i, note in enumerate(measure['data']) if note != '0']
+        events = [{'pos': e['position'], 'type': e['name'], 'value': e['value']}
+                  for e in measure['events']]
+        combined = []
+        while notes or events:
+            if events and notes:
+                if notes[0]['pos'] >= events[0]['pos']:
+                    combined.append(events.pop(0))
+                else:
+                    combined.append(notes.pop(0))
+            elif events:
+                combined.append(events.pop(0))
+            elif notes:
+                combined.append(notes.pop(0))
+
+        # Step 2: Split measure into submeasure
+        measure_cur = {'bpm': currentBPM, 'scroll': currentScroll, 'gogo': currentGogo, 'barline': currentBarline,
+                       'subdivisions': len(measure['data']), 'pos_start': 0, 'pos_end': 0,
+                       'time_sig': measure['length'], 'data': []}
+        for data in combined:
+            if data['type'] == 'note':
+                measure_cur['data'].append(data)
+                # Update the current measure's SCROLL/GOGO/BARLINE status.
+                measure_cur['scroll'] = currentScroll
+                measure_cur['gogo'] = currentGogo
+                measure_cur['barline'] = currentBarline
+                # NB: The reason we update the measure's SCROLL/GOGO/BARLINE during the "note" event is because of
+                # an ordering problem for mid-measure BPMCHANGEs. For example, imagine the following two TJA charts:
+                #   33                     11021020
+                #   #GOGOEND               #BPMCHANGE 178
+                #   #BPMCHANGE 107         #SCROLL 1.04
+                #   33,                    1102,
+                # In both examples, BPMCHANGE + one other command happen mid-measure. But, the ordering differs.
+                # This is relevant because in fumen files, "BPMCHANGE" signals the start of a new sub-measure.
+                # Yet, in both cases, we want the 2nd command to apply to the notes _after_ the BPMCHANGE.
+                # So, we make sure to only apply SCROLL/GOGO/BARLINE changes once we actually encounter new notes.
+            elif data['type'] == 'bpm':
+                currentBPM = float(data['value'])
+                # Case 1: BPM change at the start of a measure; just change BPM
+                if data['pos'] == 0:
+                    measure_cur['bpm'] = currentBPM
+                # Case 2: BPM change mid-measure, so start a new sub-measure
+                else:
+                    measure_cur['pos_end'] = data['pos']
+                    measuresCorrected.append(measure_cur)
+                    measure_cur = {'bpm': currentBPM, 'scroll': currentScroll, 'gogo': currentGogo, 'barline': currentBarline,
+                                   'subdivisions': len(measure['data']), 'pos_start': data['pos'], 'pos_end': 0,
+                                   'time_sig': measure['length'], 'data': []}
+            elif data['type'] == 'scroll':
+                currentScroll = data['value']
+            elif data['type'] == 'gogo':
+                currentGogo = bool(int(data['value']))
+            elif data['type'] == 'barline':
+                currentBarline = bool(int(data['value']))
+            else:
+                print(f"Unexpected event type: {data['type']}")
+        measure_cur['pos_end'] = len(measure['data'])
+        measuresCorrected.append(measure_cur)
+
+    return measuresCorrected
 
 
 ########################################################################################################################
